+++ src/views/MessageClients.jsx
@@ -25,6 +25,7 @@ function MessageClients(props) {
     alignContent: 'flex-start',
     maxWidth: '1600px',
     height: '100vh',
+    overflow: 'auto',
   }
 
   const summaryColumnStyles = {
@@ -94,6 +95,18 @@ function MessageClients(props) {
             Only clients considered active will be messaged. Active clients are those that are sending vitals and alerts, and have at least one
             location (button or sensor) that is sending vitals and alerts.
           </p>
+          <h4>Downtime Messages</h4>
+          <p>
+            Notice: Your Brave Buttons System is down for maintenance. During this time, Button presses will not trigger an alert. You will receive
+            another text message when everything is back online. Thank you for your patience. Have a nice day.
+          </p>
+          <p>
+            Notice: Your Brave Sensor System is down for maintenance. During this time, you may not receive bathroom alerts. You will receive another
+            text message when everything is back online. Thank you for your patience. Have a nice day!
+          </p>
+          <h4>Uptime Messasges</h4>
+          <p>Notice: Your Brave Buttons System is now back online and functioning normally. Thank you!</p>
+          <p>Notice: Your Brave Sensor System is now back online and functioning normally. Thank you!</p>
           <Form.Group>
             <Form.Label>Product</Form.Label>
             <Form.Select value={product} onChange={e => setProduct(e.target.value)}>